Resolves #5: fixes broken test of OlderCoaster. (#6)
This change to the spec file has been deployed to the solution branch of the learn-co-students repo, but not to the master branch. This requires that students edit their test file to match the one in the solution branch in order to pass all the tests. This change needs to be deployed to the master branch of the learn-co-students repo.